TPZSolutionMatrix has default ctor again
Currently there is the need to set the matrix type when creating
TPZSolutionMatrix in the TPZAnalysis classes.

This is due to the fact that there is currently no parameter in
TPZCompMesh/TPZAnalysis classes indicating whether the problem
to be solved is a complex or a real one.

The default ctor of TPZSolutionMatrix will set its type to
"Undefined" and it can be set real or complex by the assignment
operator taking TPZFMatrix<T>.