Refactor on TPZAnalysis hierarchy (see comment)
Biggest impact:
Access the TPZMatrixSolver<T> of your TPZAnalysis instance using
the method

TPZAnalysis::MatrixSolver<T>()

instead of

TPZAnalysis::Solver().

This commit began with changing the solver type of TPZAnalysis::fSolver
to TPZSolver* instead of TPZMatrixSolver<STATE>. There were
two motivations behind this: the first one is to make TPZAnalysis
suitable for both complex and real problems, and the second one was
to prepare the TPZAnalysis hierarchy for the split between the
analysis of equation systems (TPZStaticAnalysis?) and analysis
of eigenvalue systems (TPZEigenAnalysis). It makes no sense for
TPZEigenAnalysis to have a TPZMatrixSolver as its main solver.