Created TPZExactFunction + added EvaluateError for interpo. space.
The exact solution will be sent to the material instead of
the TPZAnalysis object.

This commit addresses two important aspects
of this upcoming change:

1. There was no TPZInterpolationSpace::EvaluateError(TPZVec<REAL>
&errors, bool store_error ) implementation. This new signature
is now required.

2. It was a bit complicated to create a
TPZAutoPointer<TPZFuncion<STATE>> from an existing solution that was
declared as a function. The class TPZExactFunction comes to alleviate
this issue, providing a bit more flexibility than the
TPZDummyFunction<STATE> class (that can also be used for the
same purpose). If the TPZMaterial starts storing the exact solution as a
std::function attribute, then this class is no longer necessary.

A couple of changes were performed in existing code related with error
analysis.