Introduced TPZStructMatrixT<T> + fixed a few compiler errors
To make things simpler, since every struct matrix should be a template
class, it should inherit from TPZStructMatrixT<TVar>.

The introduction of this class caused a few compiler errors, since the
compiler wasn't able to automatically deduct the template parameter
STATE in a few occasions. This was actually good, because it would be
hard to find these implicit STATE dependencies.