Setting simulation type (real/complex) when creating TPZCompMesh
This information is propagated to TPZAnalysis. When setting
solvers/struct matrices, there is a type check to ensure
compatibility.

This change has NOT been implemented in all derived classes of
TPZCompMesh/TPZAnalysis, and it will be done gradually.

p.s. A dummy complex version of TPZMaterial::Contribute was inserted
just to allow the library to compile. Since the TPZMaterial hierarchy
will be refactored, this is not a huge issue.