Refactored AutoFill methods on matrices classes
This refactor aimed to eliminate a clang warning related to usage of
RAND_MAX with some floating point types, but it was actually more
complicated since it revealed that the AutoFill methods were, in many
cases, generating only real entries even on matrices with complex
variables.

Now the AutoFill matrices are diagonally dominant and, when
appropriate, hermitian.