Changed signature of TPZCompEl::ComputeSolution
First of all, it is no longer a method from TPZCompEl, but
TPZInterpolationSpace.

Now there is only once signature that will be called:
TPZVec<REAL>&qsi, TPZMaterialData &data, bool hasPhi

If hasPhi is true, an internal implementation of ComputeSolution can
be called. If hasPhi is false, the data parameter will be assumed to
be empty and it will be accordingly initialised.

This signature is no longer virtual. Elements can then override
TPZInterpolationSpace::ReallyComputeSolution(TPZMaterialData&) to
implement different behaviour.