fix(hcurl): adds hcurl no grads compel (before it was only for hdiv)
The idea is that TPZCompElHCurl should act as an Hcurl element
even if the nograds family is selected.

In order to use the curl of the shape functions to create an hdiv
space, the derived class should be used (TPZHDivKernel3D).

Only in the derived class the additional modifications should be made.